Provide statement generator for pQualifyColumns
This is an another addition to the introduced system
of statements/procedures generators.

This addition provides a generator for procedures like

  pQualifyColumns(t tTable) returns vTable

where vTable is a view that aliases all columns of tTable
by prefixing ones by tTable and two underscores.

This procedure is an utility for a manual implementation
of custom procedures that are expected to return
fully qualified columns as shown above.