Introduce ordered StatementGeneratorProps
Drop statement generator named groups and provide each
statement generator as a call-by-name member of
a StatementGeneratorProps instance that has a description
and extends Ordered for manual initialization order
resolution.

All statements generators are returned as a Traversable
of StatementGeneratorProps, and users should take care about
using the declared statements order correctly.

The purpose of this change is to provide a way
for a manual statements initialization order resolution.